refactor controls: rename and reorganize components
- Rename ControlPanel → SelectionPanel, SelectionControls → ParameterSelection
- Rename IndicatorParameters → IndicatorFilters
- Flatten component hierarchy: move files out of ExploreControls and IndicatorSelection subfolders
- Move SelectionButton, SwapButton, SelectionPanel into controls/components/
- Move GeographySelection into controls/GeographySelection/
- Move IndicatorFilters to explore/components/
- Update all imports across routes and components
