--- src/lib/components/controls/ExploreControls/index.js
@@ -1,6 +1,0 @@
-export { default as ControlTabs } from '../ControlTabs.svelte';
-export { default as SwapButton } from './SwapButton.svelte';
-export { default as GeographySelection } from './GeographySelection/GeographySelection.svelte';
-export { default as IndicatorSelection } from './IndicatorSelection/IndicatorSelection.svelte';
-export { default as IndicatorParameters } from './IndicatorSelection/IndicatorParameters.svelte';
-export { default as SelectionControls } from './SelectionControls.svelte';
