+++ src/lib/components/controls/ModeSelectionTabs.svelte
