+++ src/lib/components/controls/components/SelectionPanel.svelte
