+++ src/routes/(default)/impacts/explore/components/ImpactGeo/SegmentedControl.svelte
