+++ src/routes/(default)/landing-page/sections/SectionExplore.svelte
@@ -3,7 +3,8 @@
   import { PATH_EXPLORE, PATH_IMPACT, URL_PATH_GEOGRAPHY, URL_PATH_INDICATOR } from '$config';
   import { IS_EMPTY_GEOGRAPHY, IS_EMPTY_INDICATOR, IS_COMBINATION_AVAILABLE_INDICATOR, CURRENT_GEOGRAPHY_UID, CURRENT_INDICATOR_UID } from '$stores/state.js';
   import { buildURL } from '$lib/utils/url.js';
-  import { SelectionControls, ControlTabs } from '$lib/components/controls/ExploreControls';
+  import ParameterSelection from '$lib/components/controls/ParameterSelection.svelte';
+  import ModeSelectionTabs from '$lib/components/controls/ModeSelectionTabs.svelte';
   import LinkArrow from '$lib/components/icons/LinkArrow.svelte';
   import Button from '$lib/components/ui/Button.svelte';
 
@@ -31,11 +32,11 @@
   <div class="md:pt-16 pb-6 md:pb-12 bg-slate-50 border-t border-contour-weakest">
     <div class="pt-8">
       <div class="mx-auto max-w-7xl px-6">
-        <ControlTabs />
+        <ModeSelectionTabs />
       </div>
     </div>
     <hr class="border-t border-contour-weakest" />
-    <SelectionControls />
+    <ParameterSelection />
     <div class="flex justify-end mt-8 px-6">
     <Button disabled={!isValidSelection} on:click={viewResults}>
       View results
